refactor(store/sql): rename rel to sql
