+++ internal/store/sql/options.go
@@ -1,4 +1,4 @@
-package rel
+package sql
 
 type (
 	Options struct {
